Unify child handling in uv crate
Removes some code duplication
